# services/timestore/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Apphub Timestore API",
    "version": "1.0.0",
    "description": "HTTP API for ingesting, querying, and managing time series datasets stored within Apphub Timestore."
  },
  "components": {
    "schemas": {
...
      "def-0": {
        "anyOf": [
          {
            "description": "Arbitrary JSON value.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              },
              {
                "type": "integer"
              },
              {
                "type": "boolean"
              },
              {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/def-0"
                }
              },
              {
                "type": "object",
                "additionalProperties": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            ]
          },
          {
            "type": "null"
          }
        ],
        "title": "https://timestore.apphub/schemas/JsonValue.json"
      },
...
      "def-2": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable description of the error."
          },
          "details": {
            "anyOf": [
              {
                "description": "Arbitrary JSON value.",
                "oneOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "integer"
                  },
                  {
                    "type": "boolean"
                  },
                  {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/def-0"
                    }
                  },
                  {
                    "type": "object",
                    "additionalProperties": {
                      "$ref": "#/components/schemas/def-0"
                    }
                  }
                ]
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "https://timestore.apphub/schemas/ErrorResponse.json"
      },
...
      "def-12": {
        "type": "object",
        "required": [
          "schema",
          "partition",
          "rows"
        ],
        "additionalProperties": false,
        "properties": {
          "datasetName": {
            "type": [
              "string",
              "null"
            ],
            "description": "Display name to assign if the dataset is created automatically."
          },
          "storageTargetId": {
            "type": [
              "string",
              "null"
            ],
            "description": "Explicit storage target identifier. Defaults to the dataset's configured target."
          },
          "tableName": {
            "type": [
              "string",
              "null"
            ],
            "description": "Physical table name override for the dataset backend."
          },
          "schema": {
            "type": "object",
            "required": [
              "fields"
            ],
            "additionalProperties": false,
            "properties": {
              "fields": {
                "type": "array",
                "minItems": 1,
                "description": "Field definitions describing the expected columns.",
                "items": {
                  "type": "object",
                  "required": [
                    "name",
                    "type"
                  ],
                  "properties": {
                    "name": {
                      "type": "string",
                      "description": "Logical column name defined by the dataset schema."
                    },
                    "type": {
                      "type": "string",
                      "description": "Logical field type used to validate incoming rows.",
                      "enum": [
                        "timestamp",
                        "string",
                        "double",
                        "integer",
                        "boolean"
                      ]
                    }
                  }
                }
              },
              "evolution": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "defaults": {
                    "type": [
                      "object",
                      "null"
                    ],
                    "additionalProperties": {
                      "$ref": "#/components/schemas/def-0"
                    }
                  },
                  "backfill": {
                    "type": [
                      "boolean",
                      "null"
                    ]
                  }
                }
              }
            }
          },
          "partition": {
            "type": "object",
            "required": [
              "key",
              "timeRange"
            ],
            "additionalProperties": false,
            "properties": {
              "key": {
                "type": "object",
                "description": "Partition key identifying the shard the data belongs to.",
                "additionalProperties": {
                  "type": "string"
                }
              },
              "attributes": {
                "type": [
                  "object",
                  "null"
                ],
                "description": "Optional attributes describing the partition.",
                "additionalProperties": {
                  "type": "string"
                }
              },
              "timeRange": {
                "type": "object",
                "required": [
                  "start",
                  "end"
                ],
                "additionalProperties": false,
                "properties": {
                  "start": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "end": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          },
          "rows": {
            "type": "array",
            "description": "Collection of rows that should be appended to the partition.",
            "items": {
              "type": "object",
              "additionalProperties": {
                "anyOf": [
                  {
                    "description": "Arbitrary JSON value.",
                    "oneOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "number"
                      },
                      {
                        "type": "integer"
                      },
                      {
                        "type": "boolean"
                      },
                      {
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/def-0"
                        }
                      },
                      {
                        "type": "object",
                        "additionalProperties": {
                          "$ref": "#/components/schemas/def-0"
                        }
                      }
                    ]
                  },
                  {
                    "type": "null"
                  }
                ]
              }
            }
          },
          "idempotencyKey": {
            "type": [
              "string",
              "null"
            ],
            "description": "Client supplied token to deduplicate ingestion attempts.",
            "maxLength": 255
          },
          "actor": {
            "type": [
              "object",
              "null"
            ],
            "required": [
              "id",
              "scopes"
            ],
            "properties": {
              "id": {
                "type": "string",
                "description": "Identifier of the actor that initiated the ingestion."
              },
              "scopes": {
                "type": "array",
                "description": "Authorization scopes granted to the actor.",
                "items": {
                  "type": "string"
                },
                "default": []
              }
            }
          }
        },
        "title": "https://timestore.apphub/schemas/DatasetIngestionRequest.json"
      },
...
      "def-18": {
        "type": "object",
        "required": [
          "mode",
          "jobId"
        ],
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "queued"
            ]
          },
          "jobId": {
            "type": "string",
            "description": "Identifier of the enqueued ingestion job."
          }
        },
        "title": "https://timestore.apphub/schemas/DatasetIngestionQueuedResponse.json"
      },
...
      "def-28": {
        "type": "object",
        "required": [
          "rows",
          "columns",
          "mode"
        ],
        "properties": {
          "rows": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": {
                "anyOf": [
                  {
                    "description": "Arbitrary JSON value.",
                    "oneOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "number"
                      },
                      {
                        "type": "integer"
                      },
                      {
                        "type": "boolean"
                      },
                      {
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/def-0"
                        }
                      },
                      {
                        "type": "object",
                        "additionalProperties": {
                          "$ref": "#/components/schemas/def-0"
                        }
                      }
                    ]
                  },
                  {
                    "type": "null"
                  }
                ]
              }
            }
          },
          "columns": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "mode": {
            "type": "string",
            "enum": [
              "raw",
              "downsampled"
            ]
          }
        },
        "title": "https://timestore.apphub/schemas/DatasetQueryResponse.json"
      },
...
    }
  },
  "paths": {
...
    "/datasets/{datasetSlug}/ingest": {
      "post": {
        "summary": "Ingest dataset partition",
        "tags": [
          "Ingestion"
        ],
        "description": "Schedules ingestion for a dataset partition. Depending on configuration the job may run inline or be enqueued.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/def-12"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "datasetSlug",
            "required": true,
            "description": "Human-readable slug uniquely identifying a dataset."
          }
        ],
        "responses": {
          "200": {
            "description": "Ingestion job accepted for processing.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-18"
                }
              }
            }
          },
          "201": {
            "description": "Ingestion completed inline.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-17"
                }
              }
            }
          },
          "202": {
            "description": "Ingestion job enqueued for asynchronous processing.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-18"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ingestion request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-2"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-2"
                }
              }
            }
          },
          "403": {
            "description": "Caller lacks permission to ingest into this dataset.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-2"
                }
              }
            }
          },
          "404": {
            "description": "Dataset not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-2"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error while scheduling ingestion.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-2"
                }
              }
            }
          }
        }
      }
    },
    "/datasets/{datasetSlug}/query": {
      "post": {
        "summary": "Execute dataset query",
        "tags": [
          "Query"
        ],
        "description": "Executes a query over the requested dataset using partition pruning and optional downsampling.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/def-27"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "datasetSlug",
            "required": true,
            "description": "Human-readable slug uniquely identifying a dataset."
          }
        ],
        "responses": {
          "200": {
            "description": "Query executed successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-28"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-2"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-2"
                }
              }
            }
          },
          "403": {
            "description": "Caller lacks permission to query this dataset.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-2"
                }
              }
            }
          },
          "404": {
            "description": "Dataset not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-2"
                }
              }
            }
          },
          "500": {
            "description": "Failed to execute query.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-2"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://127.0.0.1:4200",
      "description": "Local development server"
    }
  ],
...
}